+++ test/e2e-v2/cases/kafka/meter/docker-compose.yml
@@ -121,26 +121,5 @@ services:
       broker-b:
         condition: service_healthy
 
-  sender:
-    image: "adoptopenjdk/openjdk8:alpine-jre"
-    volumes:
-      - ./../../../java-test-service/e2e-mock-sender/target/e2e-mock-sender-2.0.0.jar:/e2e-mock-sender-2.0.0.jar
-    command: [ "java", "-jar", "/e2e-mock-sender-2.0.0.jar" ]
-    environment:
-      OAP_HOST: oap
-      OAP_GRPC_PORT: 11800
-    networks:
-      - e2e
-    ports:
-      - 9093
-    healthcheck:
-      test: ["CMD", "sh", "-c", "nc -nz 127.0.0.1 9093"]
-      interval: 5s
-      timeout: 60s
-      retries: 120
-    depends_on:
-      oap:
-        condition: service_healthy
-
 networks:
   e2e:
